Start & stop animation by clicking on the canvas
This implementation relies on the animation frame request ID as the
primary mechanism for pausing & resuming the sim. If this parameter is
defined, it is used to cancel the current frame request, effectively
stopping the simulation (as the simulation and animation loops/functions
are one and the same). If not, it is defined in `combinate()` (the main
entry point for the simulation).

This commit also defines some static parameters for documenting
pause/resume events, as well as a formal logging interface to help clean
up the simulation loop/function.