WIP: testing new model implementations based on primitive data objects
Model classes are successfully constructed and seem to match the
expected structure. References betweek them seem to be inconsistent if
not entirely missing. Currently working on exploring the generated data
tree (game.tree) -- it can be accessed by the window.game assignment
made in main() for inspection with browser developer tools.

The typescript compiler will likely throw a handful of errors, but none
are critical to the work currently being done.