Refactor 'v' from boolean to LineState[] in drawComboFrame()
`s` must be advanced to the next potentially valid state at the end of the inner
loop because doing so requires info from the evaluation of the most recent
state,but since it is often advanced by more than one combination of line
states, the last evaluated state is unknown after `s` is advanced. Saving known
valid states in `v` before `s` is advanced makes both the last known valid state
and the next possibly valid state available after the completion of the outer
loop.

> build: 72