Implement solution algorithm with fast arrays and some game logic
The primary intent of this algorithm is to be as fast as possible, and
this implementation goes to appreciable lengths to that end. It rests on
two central features to improve performance: the use of simple arrays of
primitive data types to minimize processing overhead (compared to prior
object-oriented implementations), and critical (though not complete)
game logic to reduce the scope of possible solutions. In addition,
difficulties with the in-browser debugger (caused by missing source
maps) led to a novel utilization of generator functions for debugging
and, eventually, interactive animations.

> This algorithm is a derivative of prior implementations using
> `BigInt`s to represent game states, and artifacts of the legacy model
> are evident throughout.

## Game logic

The game of **SlitherLink** is based on two general rules that determine
"valid" configurations:

 1. Cells with numbers in them must have exactly that many "filled"
    edges/lines around their border.
 2. The solution must form a single continuous, non-intersecting path

The algorithm checks individual board states for validity as determined
by the second rule. The logic is enforced on a node-by-node basis in
terms of the number of filled/empty lines meeting at a given node.
Specifically, a node is considered valid if and only if it has exactly
zero or two filled lines; any other combination of line states would
indicate a violation of the second rule.

## Implementation details

As mentioned, this commit features the utilization of a generator
function for development/debugging purposes. This section describes the
procedure that would be followed were it not for those "interruptions".

The procedure is designed around the use of a continuous loop for
checking possible solutions, to make the most of available processing
capacity. This loop is run in intervals of fixed duration (currently
50ms) individually scheduled using `window.setTimeout()`; This is done
to balance performance with UX/UI responsiveness since, currently, the
simulation runs in the main thread and blocks the window's event loop.

The method `combinate()` initializes the loop, defining parameters that
help control execution and making the first call to `setTimeout()`. The
loop is executed in `drawComboFrame()` where, after some persistent
variables are defined, the outermost `while` loop stops checking
solutions after the given time interval and schedules the next "run".
Within this outer loop, a nested `while` loop runs the core algorithm
logic repeatedly until a valid solution is identified.

#### Procedure

Starting with the highest-order line, nodes at each line are checked for
validity as described above in *Game logic*. Even if the first node is
invalid, both nodes are always checked in order to identify the
highest-order line that can be incremented in order to skip over
implicitly invalid states (however, see notes in *Drawbacks* below).

When a given state is determined to be invalid, the lowest-index line at
the invalid node(s) is recorded to be advanced; if both nodes are
invalid, their respective lowest-order lines are compared and the
higher-order of the two is chosen. The state is then "incremented" by
adding 1 to the designated line's state and propagating results upward
to higher-order lines; lower-order lines are reset to 0 so that no
possibly valid states are skipped. The propagation process works by
treating each line in the lines array as a binary bit -- with empty and
filled states represented by `0` and `1` -- of an order equal to that
line's position in the array.

Three arrays are derived from the game board structure: one of nodes,
one of lines, and one of line states. This allows all object states and
references to be expressed in lightweight containers and accessed
locally and quickly, standing in stark contrast to the bulky, complex
object-oriented approach of the previous implementation.

The lines array, `l`, contains sub-arrays of 2 integers. Similarly, the
nodes array contains sub-arrays of 3 integers. Values in `l` are the
indices, in `n`, of that line's nodes; and values in `n` are the
indices, in `l`, of the lines that meet at that node (`-1` is assigned
in place of a non-existent line, since some nodes only have 2 lines).
The states array, `s`, is a flat array of integers, `0` or `1`,
indicating the state (empty or filled) of the line at the corresponding
position in `l`.

**prevent checking nodes redundantly**
**jsperf results for `Math.min()`**

**reset higher of two lowest-index lines**
**iterate lines in reverse order**

**game logic**

**generators for debugging/animation**

Drawbacks
- includes multi-loop solutions
- does not account for cell numbering
- can only represent 2 line states
- sometime increments lower-order lines

> build: 146