Find single-loop solutions very fast
In addition to skipping states based on possible configurations at
individual nodes, this commit further reduces the scope of possible
solutions by limiting next-line-to-be-incremented candidates to a small
number of lines (at path endpoints), and the resulting performance
improvements make this implementation faster than the previous by orders
of magnitude. On a pwerful computer, this has made it possible to find
all solutions on boards with radius=2 in approximately 30 seconds.