Check a line's nodes' DoF when asserting its state
When a line's state is asserted, either by filling it or emptying it
(even if it was already asserted), check the number of degrees of
freedom at each of its nodes and, if exactly two are asserted, assert
the third so that total number of filled lines is 0 or 2 (the rule for
node validity). This procedure ignores nodes with more or less than 2
asserted lines (accounting for the clicked line's new state).

This implementation will only initiate DoF checks when warranted. In
particular, it will *not* attemt checks for a node at which the clicked
line is the only asserted line, nor when a line state is changed
(toggled) at a node where both opposing lines are also asserted. Since a
node may always be valid when at least one of its lines is not asserted,
de-asserting a line never triggers DoF checks.

Similarly, the board will be redrawn only when a click has an observable
effect (i.e. de-asserting a line was was not asserted will not force the
board to be redrawn).