Implement prelim. incremental solution generation
Click on the SVG to run `stepSolution` & increment the solution by 1 step. Automate this process by un-commenting the `while(step)` loop (`SVGGameBoard.vue` line 115) and disabling the `<svg>` element's `@click` event handler (same file line 162).

Alternatively, sort cells manually by clicking to add to pool (only cells in the `bank`, colored tan, can be added at each step). The initial cell is selected randomly but this can be changed using `startId` (`SVGGameBoard.vue` line 93).

When a cell is added to the `pool`, its neighbors are added to a `bank`. A cell added to the `bank` multiple times before being added to the `pool` is disqualified from the `pool` (exact number of additions, or "banks", allowed is set/evaluated in `SVGGameBoard.vue` line 149). Setting this threshold above 2 may result in arbitrarily large "blobs" in the solution. Values of 1 and 2 both have pros and cons, but as of now 2 has slight preference as it allows for small blobs that make the solution shape more interesting.

Of note, this solution-generating algorithm is built around an implementation of Fisher-Yates shuffle (`https://bost.ocks.org/mike/shuffle/`, also archived) that allows the shuffled array to be updated (add/remove values) after each step in the iteration (`GameState.ts` line 56). The implementation is modified to allow additions to the array while still running in roughly O(n) time/space (assuming enough array capacity allocated initially): iteration is performed with ascending index, and excluded values are not technically removed from the original array but simply added to a second which is checked during each iteration.