Implement prelim. data entry UI component
Interface will iterate over each cell in an SVG game board and prompt for respective counts/values. Intersecting SVG lines will indicate the current row/column (in grid coordinates).

This commit implements a rudimentary system for uploading an image (screenshot of an actual game/puzzle from the Android app) onto which the SVG board will be superimposed. Four datums, defined by the user, can be used to derive a mapping of grid coordinates to screen coordinates aligned with the image.