Demo use of SVG transforms for pan/zoom
Pan offset is defined in terms of relative coords, so the translation is applied first, then scaling (SVG transforms performed in reverse/right-to-left order as listed in `transform` attribute). As a result, during panning mouse coordinate deltas must be scaled manually before being added to the current offset. The alternative -- pan offset defined in terms of screen coordinates -- would require that offsets be scaled each time the zoom level is changed (may consider doing this...).

1. Define board structure/elements (cells, lines, verts) in relative (u,v) coordinates
2. Apply SVG transform to map to view box coordinates
    - Set view box bounds to desired relative coord. bounds to make that the default/identity transform (scale = 1, translation = <0,0>)
3. Results intrinsically converted to screen coordinates