Demo use of SVG transforms for individual elements
Doing this simplifies math/geometry considerably but at the cost of tough choices on document structure. For example: positioning lines relative to cells completely eliminates the need to calculate start/end coords for individual lines, but means that most lines on the board will be duplicated as they are defined by each cell that they border. This in turn would complicate functional aspects, such as click event handlers, as well as aesthetic ones, like the use of opacity resulting in different appearance for lines around the perimeter since they *don't* have duplicates.